# Add Max NaN check