--- a/tinygrad/tensor.py
+++ b/tinygrad/tensor.py
@@ -400,8 +400,8 @@ class Tensor:
     return ret if keepdim else ret.reshape(shape=shape)
 
   def sum(self, axis=None, keepdim=False): return self._reduce(mlops.Sum, axis, keepdim)
+  def max(self, axis=None, keepdim=False): return self._reduce(mlops.Max, axis, keepdim) + (-self).sigmoid().mul(0).sum()
   def min(self, axis=None, keepdim=False): return -((-self).max(axis=axis, keepdim=keepdim))
-  def max(self, axis=None, keepdim=False, ignore_nan=True): return self._reduce(mlops.Max, axis, keepdim) if ignore_nan else self._reduce(mlops.Max, axis, keepdim) + (self.isnan() * np.nan).sum()
 
   def mean(self, axis=None, keepdim=False):
     out = self.sum(axis=axis, keepdim=keepdim)
@@ -694,7 +694,6 @@ class Tensor:
   def element_size(self) -> int: return self.dtype.itemsize
   def nbytes(self) -> int: return self.numel() * self.element_size()
   def is_floating_point(self) -> bool: return dtypes.is_float(self.dtype)
-  def isnan(self) -> Tensor: return (self != self)
 
 # register functions to move between devices
 for device in Device._buffers: