Fixes #862 
### Changes:
* adds nan check to Tensor.max()
* adds max nan CI testing
* ups kernel limits
* handles -inf edge cases

### How it works:
```
t1 = Tensor([float("nan"), float("-inf"), 5, -10])

t1 = -t1                       -> array([nan, inf, -5., 10.], dtype=float32)
t1 = t1.sigmoid()              -> array([nan,  1.,  1.,  1.], dtype=float32)
t1 = t1.mul(0)                 -> array([nan,  0.,  0.,  0.], dtype=float32)
t1 = t1.sum()                  -> array(nan, dtype=float32)
```
### Why:
I had a simpler algorithm but it was causing onnx  test_efficientnet to guess the images wrong. This was because -inf was being passed in and then converted to nan. This solves that.

### Benchmark
**Master:**
`max  4096x 4096 3.71 ms (4.52 GFLOPS 18.09 GB/s) in torch,  8.73 ms (1.92 GFLOPS 7.68 GB/s) in tinygrad,   2.35x slower 16.78 MOPS  67.11 MB`

**This Patch:**
`max 4096x 4096 3.79 ms (39.87 GFLOPS  35.44 GB/s) in torch, 17.35 ms ( 8.70 GFLOPS 7.74 GB/s) in tinygrad, 4.58x slower 150.99 MOPS 134.22 MB`

**Torch vs Tiny**
```
t1 = torch.tensor([np.nan, 1])
t1.max().numpy() -> array(nan, dtype=float32)
t2 = Tensor([np.nan, 1])
t2.max().numpy() -> array(nan, dtype=float32)
```

I understand that increasing kernel limits is probably undesirable.