@tinyb0t 2023-08-03
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/tensor.py                         1               1
------------------------------------------------------------
lines added in the tinygrad folder: 0
```