@geohot 2023-08-08
Big perf hit, I'm not going to merge this.